+++ utl/src/lib.rs
@@ -1,0 +1,19 @@
+extern crate walkdir;
+use std::io;
+use std::path::Path;
+use std::path::PathBuf;
+use walkdir::WalkDir;
+
+pub fn list_files(directory: &Path, file_ext: &str) -> io::Result<Vec<PathBuf>> {
+    let mut list: Vec<PathBuf> = Vec::new();
+
+    for entry in WalkDir::new(directory) {
+        let entry = entry.unwrap();
+        let path = entry.path().to_path_buf();
+        if path.is_file() && path.to_str().unwrap().ends_with(file_ext) {
+            //println!("list_files: {}", entry.path().display());
+            list.push(path);
+        }
+    }
+    Ok(list)
+}