+++ vq/src/lib.rs
@@ -1,12 +1,14 @@
 extern crate libc;
 extern crate structopt;
+extern crate utl;
+
+use std::ffi::CString;
+use std::path::Path;
+use std::path::PathBuf;
 
 use libc::c_char;
 use libc::c_double;
 use libc::c_int;
-use std::ffi::CString;
-use std::path::PathBuf;
-
 use structopt::StructOpt;
 
 extern "C" {
@@ -33,7 +35,8 @@ pub struct VqLearnOpts {
     #[structopt(short = "w", long = "class-name")]
     class_name: Option<String>,
 
-    /// Predictor files for training
+    /// Predictor files for training. If a directory is given, then
+    /// all `.prd` under it will be used.
     #[structopt(parse(from_os_str))]
     predictor_filenames: Vec<PathBuf>,
 }
@@ -52,9 +55,24 @@ pub fn main_vq_learn(opts: VqLearnOpts) {
     })
     .unwrap();
 
-    let num_predictors = predictor_filenames.len() as c_int;
+    let actual_predictor_filenames: Vec<PathBuf> = if predictor_filenames.len() == 1 {
+        let path = Path::new(&predictor_filenames[0]);
+        if path.is_dir() {
+            match utl::list_files(path, "prd") {
+                Ok(list) => list,
+                Err(err) => panic!("cannot list dir {}: {}", path.display(), err),
+            }
+        } else {
+            predictor_filenames
+        }
+    } else {
+        predictor_filenames
+    };
 
-    let c_strings: Vec<CString> = predictor_filenames
+    let num_actual_predictors = actual_predictor_filenames.len() as c_int;
+    println!("num_actual_predictors: {}", num_actual_predictors);
+
+    let c_strings: Vec<CString> = actual_predictor_filenames
         .into_iter()
         .map(|predictor_filename| {
             let str = predictor_filename.to_str().unwrap();
@@ -79,7 +97,7 @@ pub fn main_vq_learn(opts: VqLearnOpts) {
             epsilon as c_double,
             codebook_class_name.as_ptr() as *const i8,
             c_predictor_filenames.as_ptr(),
-            num_predictors,
+            num_actual_predictors,
         );
     }
 }